combine all under one survey
