+++ client/src/App.js
@@ -3,8 +3,7 @@ import './App.css';
 import Landing from './components/ss_landing/landing';
 import Homepage from './components/ss_homepage/homepage';
 import Register from './components/ss_register/register';
-import SurveyMeal from './components/ss_surveymeal/surveymeal';
-import SurveyWorkout from './components/ss_surveyworkout/surveyworkout';
+import Survey from './components/ss_survey/survey';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 
 function App() {
@@ -14,8 +13,7 @@ function App() {
         <Route path="/" element={<Landing />} />
         <Route path="/homepage" element={<Homepage />} />
         <Route path="/register" element={<Register />} />
-        <Route path="/surveymeal" element={<SurveyMeal />} />
-        <Route path="/surveyworkout" element={<SurveyWorkout />} />
+        <Route path="/survey" element={<Survey />} />
       </Routes>
     </BrowserRouter>
   );
