+++ client/src/components/ss_landing/landing.js
@@ -32,7 +32,7 @@ const Landing = () => {
 
           if (response.ok) {
               localStorage.setItem('userId', data.userId);
-              navigate(data.firstTimeLogin ? '/surveymeal' : '/homepage');
+              navigate(data.firstTimeLogin ? '/survey' : '/homepage');
           } else {
               alert(data.message);
           }
