--- client/src/components/ss_surveyworkout/surveyworkout.css
@@ -1,43 +1,0 @@
-.question{
-    display: none;
-}
-
-.page {
-    display: flex;
-    height: 100vh;
-    width: 100vw;
-    justify-content: center;
-    align-items: center;
-}
-
-.intake-survey {
-    display: flex;
-    height: 256px;
-    width: 512px;
-    flex-direction: column;
-    justify-content: space-between;
-    align-items: center;
-    position: absolute;
-    margin: 10px;
-    border: 5px solid rgba(255, 255, 255, 0.65);
-    border-radius: 20px;
-    background-image: linear-gradient(to bottom, rgb(116, 179, 195), rgb(22, 76, 120));
-}
-
-.question-buttons {
-    display: flex;
-    width: 250px;
-    justify-content: space-evenly;
-    align-items: center;
-    flex-direction: row;
-}
-
-button {
-    height: 25px;
-    width: 75px;
-    background-color: rgba(255, 255, 255, 0.65);
-    border: 2px solid grey;
-    border-radius: 20px;
-    color: black;
-    font-weight: 1000;
-}
